Rename simulated test/network
