+++ radixdlt/src/integration/java/com/radixdlt/consensus/simulation/bft/synchronous/FPlusOneOutOfBoundsTest.java
@@ -20,8 +20,8 @@ package com.radixdlt.consensus.simulation.bft.synchronous;
 import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
 
 import com.radixdlt.consensus.simulation.BFTCheck.BFTCheckError;
-import com.radixdlt.consensus.simulation.BFTSimulatedTest;
-import com.radixdlt.consensus.simulation.BFTSimulatedTest.Builder;
+import com.radixdlt.consensus.simulation.SimulatedTest;
+import com.radixdlt.consensus.simulation.SimulatedTest.Builder;
 import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.TimeUnit;
@@ -31,7 +31,7 @@ public class FPlusOneOutOfBoundsTest {
 	private final int latency = 50;
 	private final int synchronousTimeout = 8 * latency;
 	private final int outOfBoundsLatency = synchronousTimeout;
-	private final Builder bftTestBuilder = BFTSimulatedTest.builder()
+	private final Builder bftTestBuilder = SimulatedTest.builder()
 		.pacemakerTimeout(2 * synchronousTimeout)
 		.checkSafety("safety")
 		.checkNoneCommitted("noneCommitted");
@@ -41,7 +41,7 @@ public class FPlusOneOutOfBoundsTest {
 	 */
 	@Test
 	public void given_0_out_of_3_nodes_out_of_synchrony_bounds() {
-		BFTSimulatedTest test = bftTestBuilder
+		SimulatedTest test = bftTestBuilder
 			.numNodesAndLatencies(3, latency, latency, latency)
 			.build();
 
@@ -54,7 +54,7 @@ public class FPlusOneOutOfBoundsTest {
 	 */
 	@Test
 	public void given_1_out_of_3_nodes_out_of_synchrony_bounds() {
-		BFTSimulatedTest test = bftTestBuilder
+		SimulatedTest test = bftTestBuilder
 			.numNodesAndLatencies(3, latency, latency, outOfBoundsLatency)
 			.build();
