+++ radixdlt/src/integration/java/com/radixdlt/consensus/simulation/checks/AllProposalsHaveDirectParentsCheck.java
@@ -21,7 +21,7 @@ import com.radixdlt.consensus.simulation.BFTCheck;
 import com.radixdlt.consensus.EventCoordinatorNetworkRx;
 import com.radixdlt.consensus.Proposal;
 import com.radixdlt.consensus.Vertex;
-import com.radixdlt.consensus.simulation.SimulatedBFTNetwork.RunningNetwork;
+import com.radixdlt.consensus.simulation.SimulatedNetwork.RunningNetwork;
 import com.radixdlt.crypto.ECKeyPair;
 import io.reactivex.rxjava3.core.Observable;
 import java.util.List;
